v5.x-scala-js: mapFirst, mapSecond moved to sdk, IntegralFromExactIntegral
