+++ core-lib/shared/src/main/scala/special/collection/Extensions.scala
@@ -1,7 +1,6 @@
 package special.collection
 
 import debox.cfor
-import scalan.RType
 
 object Extensions {
   implicit class CollOps[T](val coll: Coll[T]) extends AnyVal {
@@ -14,9 +13,6 @@ object Extensions {
   }
 
   implicit class PairCollOps[A,B](val source: Coll[(A,B)]) extends AnyVal {
-    @inline def mapFirst[A1: RType](f: A => A1): Coll[(A1, B)] = source.asInstanceOf[PairColl[A,B]].mapFirst(f)
-    @inline def mapSecond[B1: RType](f: B => B1): Coll[(A, B1)] = source.asInstanceOf[PairColl[A,B]].mapSecond(f)
-
     def foreach(f: (A, B) => Unit) = {
       val b = source.builder
       val (as, bs) = b.unzip(source)
